test: share the canopy view mock between suites
