+++ tests/meridian-client.test.ts
@@ -1,31 +1,7 @@
-import { jest } from "@jest/globals";
 import { normalizeMoveAddress } from "../packages/core/src";
 import { CanopySdk } from "../packages/sdk/src";
 import { DEFAULT_VIEW_BATCH_SIZE } from "../packages/sdk/src/internal/address-batches";
-
-interface MockViewClient {
-  view: jest.MockedFunction<(input: unknown) => Promise<unknown[]>>;
-}
-
-function createMovementMock(
-  responses: Record<string, unknown[] | ((args: unknown[]) => unknown[])>
-): MockViewClient {
-  return {
-    view: jest.fn(async (input: unknown) => {
-      const payload = (input as { payload: { function: string; functionArguments?: unknown[] } })
-        .payload;
-      const response = responses[payload.function];
-
-      if (!response) {
-        throw new Error(`Missing mock response for ${payload.function}`);
-      }
-
-      return typeof response === "function"
-        ? response(payload.functionArguments ?? [])
-        : response;
-    }),
-  };
-}
+import { createMovementMock } from "./fixtures/view-client-mock";
 
 describe("Meridian client", () => {
   it("dedupes and chunks meridian batch view requests", async () => {
